add signature function

+++ src/main/java/com/versionsystem/service/impl/drug/WDrugListService.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.versionsystem.common.ApplicationError;
+import com.versionsystem.common.ApplicationErrorCode;
 import com.versionsystem.common.ApplicationException;
 import com.versionsystem.common.ApplicationParas;
 import com.versionsystem.common.BeanUtils;
@@ -43,8 +44,7 @@ public class WDrugListService {
 		for(WDrugList o:list){
 			vo=new WDrugListUI();
 			BeanUtils.copyProperties(o,vo);
-			//String jsonDateFormat=configService.getProperty(ApplicationParas.APP_JSONDATEFORMAT);
-			//vo.setEffectiveFrom(ObjectConverter.toUTCDate2(o.getEffectiveFrom(),jsonDateFormat));
+			
 			rl.add(vo);
 		}
 		return rl;
@@ -57,9 +57,15 @@ public class WDrugListService {
 	public boolean delete(long id){
 		WDrugList o=this.drugListRepository.findOne(id);
 		if(o!=null){
-			this.drugListRepository.delete(o);
-			logger.info(this.userService.getCurrentUser()+" delete a  record drug code="+o.getDrugName()+" on "+new Date());
-			return true;
+			String errorCode=this.validateBeforeDelete(id);
+			if(errorCode==null){
+				this.drugListRepository.delete(o);		
+				logger.info(this.userService.getCurrentUser()+" delete a  record drug code="+o.getDrugName()+" on "+new Date());
+				return true;
+			}
+			else
+				throw new ApplicationException(new ApplicationErrorCode(errorCode,"Validation Error").toString());
+				
 		}else
 			throw new ApplicationException(ApplicationError.ObjectNotFoundException.toString());
 			
@@ -67,11 +73,16 @@ public class WDrugListService {
 	
 	@Transactional
 	public long create(WDrugListUI temp){
-				
+		String errorCode=this.validateBeforeSave(temp);
+		if(errorCode==null){
 				WDrugList o=new WDrugList();
 				BeanUtils.copyProperties(temp, o);
 				WDrugList saved=this.drugListRepository.saveAndFlush(o);
-				return saved.getId();					
+				return saved.getId();	
+		}
+		else{
+			throw new ApplicationException(new ApplicationErrorCode(errorCode,"Validation Error").toString());
+		}
 		
 	}
 	
@@ -83,16 +94,41 @@ public class WDrugListService {
 			if(ci==null)
 				throw new ApplicationException(ApplicationError.ObjectNotFoundException.toString()+" WDrugList"+temp.getId()+" not found");
 			else{
-				WDrugList o=this.drugListRepository.findOne(temp.getId());
-				BeanUtils.copyProperties(temp, o);
-				
-				this.drugListRepository.saveAndFlush(o);
-				return true;
+				String errorCode=this.validateBeforeUpdate(temp);
+				if(errorCode==null){
+					WDrugList o=this.drugListRepository.findOne(temp.getId());
+					BeanUtils.copyProperties(temp, o);			
+					this.drugListRepository.saveAndFlush(o);
+					return true;
+				}
+				else
+					throw new ApplicationException(new ApplicationErrorCode(errorCode,"Validation Error").toString());
 			}
 		
 				
 		
 	}
+	
+	//返回空表示可以继续下一步动作，否则，表示有错误产生，不可下一步动作，这里可以验证表单中的某些值是否符合特定的业务逻辑
+	//若有错，就返回错误编码
+	private String validateBeforeSave(WDrugListUI vo){
+		
+		return null;
+	}
+	
+	//返回空表示可以继续下一步动作，否则，表示有错误产生，不可下一步动作，这里可以验证表单中的某些值是否符合特定的业务逻辑
+	//若有错，就返回错误编码
+	private String validateBeforeUpdate(WDrugListUI vo){
+			
+		return null;
+	}
+	
+	//返回空表示可以继续下一步动作，否则，表示有错误产生，不可下一步动作，这里可以验证表单中的某些值是否符合特定的业务逻辑
+	//若有错，就返回错误编码
+	private String validateBeforeDelete(long id){
+				
+		return null;
+	}
